+++ Sources/MapboxDirections/Directions.swift
@@ -110,7 +110,7 @@ open class Directions: NSObject {
     /**
      The shared directions object.
      
-     To use this object, a Mapbox [access token](https://docs.mapbox.com/help/glossary/access-token/) should be specified in the `MGLMapboxAccessToken` key in the main application bundle’s Info.plist.
+     To use this object, a Mapbox [access token](https://docs.mapbox.com/help/glossary/access-token/) should be specified in the `MBXAccessToken` key in the main application bundle’s Info.plist.
      */
     public static let shared = Directions()